+++ src/usecases/implementations/user/update-user-social-media/update-user-social-media-usecase.ts
@@ -1,8 +1,9 @@
 import { type UpdateUserSocialMediaData, type UpdateUserSocialMedia, type UpdateUserSocialMediaResponse } from '@/domain/contracts/user'
 import { InvalidSocialMediaIdError } from '@/domain/errors'
 import { left, right } from '@/shared/either'
-import { type AddOrUpdateUserSocialMediaByIdsRepo } from '@/usecases/contracts/db/social-media/add-or-update-user-social-media-by-ids-repo'
+
 import { type FindSocialMediaByIdRepo } from '@/usecases/contracts/db/social-media/find-social-media-by-id-repo'
+import { type AddOrUpdateUserSocialMediaByIdsRepo } from '@/usecases/contracts/db/user/add-or-update-user-social-media-by-ids-repo'
 
 export class UpdateUserSocialMediaUseCase implements UpdateUserSocialMedia {
   constructor (
